feat: cleanup-ui-activity-card-coverage-split-002 - Preserve graph projection and selection semantics in focused suites
