+++ ui/biome.jsonc
@@ -189,6 +189,7 @@
         "src/features/workflow-activity/components/react-flow-current-activity-card-import.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card-editor-chrome.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card-import-flows.test.tsx",
+        "src/features/workflow-activity/components/react-flow-current-activity-card-graph-semantics.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card-component.harness.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card.coverage.test.tsx",
         "src/features/workflow-activity/components/react-flow-current-activity-card.test.tsx",
